added better exception handlers

+++ src/main/java/com/atwo/paganois/services/VerificationService.java
@@ -1,6 +1,7 @@
 package com.atwo.paganois.services;
 
 import java.time.LocalDateTime;
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +11,9 @@ import org.springframework.stereotype.Service;
 import com.atwo.paganois.entities.TokenType;
 import com.atwo.paganois.entities.User;
 import com.atwo.paganois.entities.VerificationToken;
+import com.atwo.paganois.exceptions.ExpiredTokenException;
+import com.atwo.paganois.exceptions.InvalidTokenTypeException;
+import com.atwo.paganois.exceptions.TokenNotFoundException;
 import com.atwo.paganois.repositories.VerificationTokenRepository;
 
 import jakarta.transaction.Transactional;
@@ -31,13 +35,14 @@ public class VerificationService {
 
     @Transactional
     public void sendPasswordReset(String email) {
-        // TODO: adicionar e tratar exceptions (MessagingException)
-        User user = userDetailsService.findByEmail(email);
+        Optional<User> userOptional = userDetailsService.findByEmailOptional(email);
+        if (userOptional.isEmpty())
+            return;
 
-        // Remove tokens antigos
+        User user = userOptional.get();
+
         tokenRepository.deleteByUserIdAndType(user.getId(), TokenType.PASSWORD_RESET);
 
-        // Gera novo token
         String token = UUID.randomUUID().toString();
 
         VerificationToken resetToken = new VerificationToken();
@@ -47,13 +52,9 @@ public class VerificationService {
         resetToken.setExpiryDate(LocalDateTime.now().plusHours(1));
         tokenRepository.save(resetToken);
 
-        // Monta URL de reset
         String resetUrl = baseUrl + "/auth/reset-password?token=" + token;
 
-        // Envia email
-        emailService.sendSimpleEmail(
-                email,
-                "Resetar senha - Paganois",
+        emailService.sendSimpleEmail(email, "Resetar senha - Paganois",
                 "Clique no link a seguir para resetar sua senha: \n" + resetUrl
                         + "\n PS: testar via API Clients (Postman, Insonmia) caso sem frontend");
     }
@@ -80,23 +81,19 @@ public class VerificationService {
         String confirmationUrl = baseUrl + "/auth/verify-email?token=" + token;
 
         // Envia email
-        emailService.sendSimpleEmail(
-                user.getEmail(),
-                "Confirme seu email - Paganois",
+        emailService.sendSimpleEmail(user.getEmail(), "Confirme seu email - Paganois",
                 "Por favor, confirme seu email clicando no link: \n" + confirmationUrl);
     }
 
-    // TODO: tratar as excessões (e especializar)
     @Transactional
     public VerificationToken validateToken(String token, TokenType type) {
-        VerificationToken tokenEntity = tokenRepository
-                .findByToken(token)
-                .orElseThrow(() -> new RuntimeException("Token inválido"));
+        VerificationToken tokenEntity = tokenRepository.findByToken(token)
+                .orElseThrow(() -> new TokenNotFoundException("Token não encontrado"));
         if (tokenEntity.getType() != type)
-            throw new RuntimeException("Token inválido");
+            throw new InvalidTokenTypeException("Token com tipo inválido");
 
         if (tokenEntity.isExpired()) {
-            throw new RuntimeException("Token expirado");
+            throw new ExpiredTokenException("Token expirado");
         }
 
         return tokenEntity;
